requirements.txt
azure-identity
azure-mgmt-storage
azure-storage-blob
python-dotenv
